Replace default collection grid in favor of a custom section
The new section is heavily inspired by the default one. It uses some
of its orginal settings, adds new ones (i.e. `show_out_of_stock`),
and occasionally uses hardcoded values (product card padding) just
for variety.

When possible, functionalities are added by reusing features from
the existing template.
The product-card snippet has been modified as well in order add the
new features, which are driven by flags, so changes are backwards
compatible.

The product metadata information is called `donation_amount`, it
represents the amount of the product price that is donated to OSS
projects. I tried with a type Money metadata field, but it was
quite hard to display, actually I found no way to do that, so I
added this other field (decimal) that I was able to make work.